Move all scripts to TS, split end.js into multiple files
