+++ .gitignore
@@ -1,4 +1,5 @@
 ./vscode/ipch/*
+.ignored/**/*
 out
 node_modules
 build
