+++ manifests/debug.json
@@ -1,12 +1,12 @@
 {
     "name": "ProgTest Themes",
-    "version": "100.0.32",
+    "version": "100.0.33",
     "description": "Theme manager for ProgTest",
     "manifest_version": 2,
     "content_scripts": [
         {
             "run_at": "document_start",
-            "js": ["content/start.js", "content/highlightjs/highlight.pack.js"],
+            "js": ["content/start.js", "content/highlightjs/highlight.min.js"],
             "matches": ["*://progtest.fit.cvut.cz/*", "*://ptmock.localhost/*"],
             "all_frames": true
         },
@@ -38,5 +38,11 @@
     "icons": {
         "128": "icon.png"
     },
-    "web_accessible_resources": ["themes/*"]
+    "web_accessible_resources": ["themes/*"],
+    "applications": {
+        "gecko": {
+            "id": "progtest-themes@debug.dev",
+            "update_url": "https://raw.githubusercontent.com/keombre/progtest-theme/stable/update.json"
+        }
+    }
 }
