+++ manifests/firefox.json
@@ -6,7 +6,7 @@
     "content_scripts": [
         {
             "run_at": "document_start",
-            "js": ["content/start.js", "content/highlightjs/highlight.pack.js"],
+            "js": ["content/start.js", "content/highlightjs/highlight.min.js"],
             "matches": ["*://progtest.fit.cvut.cz/*"],
             "all_frames": true
         },
@@ -17,10 +17,7 @@
             "all_frames": true
         }
     ],
-    "background": {
-        "scripts": ["background.js"],
-        "persistent": true
-    },
+    "background": { "scripts": ["background.js"], "persistent": true },
     "permissions": [
         "webRequest",
         "webRequestBlocking",
@@ -34,9 +31,7 @@
         "default_icon": "icon.png",
         "default_popup": "options/options.html"
     },
-    "icons": {
-        "128": "icon.png"
-    },
+    "icons": { "128": "icon.png" },
     "web_accessible_resources": ["themes/*"],
     "applications": {
         "gecko": {
