+++ scripts/web-ext.d.ts
@@ -1,9 +1,8 @@
-/// <reference types="web-ext" />
 declare module "web-ext" {
     export const cmd: {
         [K in keyof WebExtConfig]: (
             options: WebExtSharedOptions & WebExtConfig[K],
-            { shouldExitProgram }: { shouldExitProgram: boolean }
+            { shouldExitProgram }: { shouldExitProgram: boolean },
         ) => Promise<void>;
     };
