+++ src/messages.ts
@@ -1,0 +1,3 @@
+export enum MessageType {
+    GET_SETTINGS = "GET_SETTINGS",
+}
