+++ src/themes/automatic.css
@@ -1,2 +1,3 @@
-@import url("./light.css") screen and (prefers-color-scheme: light);
+@import url("./light.css") screen and not (prefers-color-scheme: dark);
 @import url("./dark.css") screen and (prefers-color-scheme: dark);
+@import url("./common.css");
