+++ src/themes/common.css
@@ -297,7 +297,9 @@ select.std.w200[name="UID_UNIVERSITY"] {
     z-index: 2;
     width: 250px;
     text-align: left;
-    transition: font-size 0.2s ease, top 0.2s ease;
+    transition:
+        font-size 0.2s ease,
+        top 0.2s ease;
     padding: 1px 0;
 }
 
@@ -526,7 +528,7 @@ div.subjectSelect .updButton {
     padding: 2px 5px;
 }
 
-div.subjectSelect > a {
+div.subjectSelect > .subject {
     border-radius: 7px;
     padding: 16px 12pt;
     min-height: 140px;
@@ -552,22 +554,22 @@ div.subjectSelect.mainInfo > a {
     }
 }
 
-div.subjectSelect > a:hover {
+div.subjectSelect > .subject:hover {
     box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.25);
 }
 
-div.subjectSelect > a > div.subject-title {
+div.subjectSelect > .subject > div.subject-title {
     font-size: 16pt;
     width: calc(100% - 80px);
 }
 
-div.subjectSelect > a > div.subject-body {
+div.subjectSelect > .subject > div.subject-body {
     font-size: 12pt;
     width: calc(100% - 85px);
     margin-top: 5px;
 }
 
-div.subjectSelect > a > div.subject-footer {
+div.subjectSelect > .subject > div.subject-footer {
     position: absolute;
     bottom: 0;
     left: 0;
